parsing requests, storing captures in db

+++ app/models/Capture.scala
@@ -2,13 +2,13 @@ package models
 
 import com.mongodb.casbah.Imports._
 
-case class Capture(id: Long, x: Integer, y: Integer)
+case class Capture(id: Long, x: Double, y: Double)
 
 object Capture {
   val client = MongoClient()
   val collection = client("segments")("captures")
 
-  def create(id: Long, x: Integer, y: Integer) {
+  def create(id: Long, x: Double, y: Double) {
     val capture = MongoDBObject("id" -> id, "x" -> x, "y" -> y)
     collection.insert(capture)
   }
